Modification de logiques de classe, début des test avec le command runner

+++ src/main/java/com/example/demo/models/Member.java
@@ -22,9 +22,9 @@ public class Member {
     private String firstName;
     private String lastName;
     private String email;
-    private Long pswd;
+    private int pswd;
 
-    public Member(String firstName, String lastName, String email, Long pswd){
+    public Member(String firstName, String lastName, String email, int pswd){
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
